bigquery-function/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
pyarrow
